+++ src/Migration/Manager.php
@@ -73,7 +73,7 @@ class Manager
     {
         $filesFinder = new FilesFinder();
         foreach ($this->config->getMigrationDirs() as $identifier => $directory) {
-            if (empty($dirs) || (!empty($dirs) && in_array($identifier, $dirs))) {
+            if (empty($dirs) || (!empty($dirs) && in_array($identifier, $dirs, true))) {
                 $filesFinder->addDirectory($directory);
             }
         }
@@ -83,7 +83,7 @@ class Manager
             require_once $file;
             $classNameCreator = new ClassNameCreator($file);
             $className = $classNameCreator->getClassName();
-            if (empty($classes) || (!empty($classes) && in_array($className, $classes))) {
+            if (empty($classes) || (!empty($classes) && in_array($className, $classes, true))) {
                 $migrationIdentifier = $classNameCreator->getDatetime() . '|' . $className;
                 $migrations[$migrationIdentifier] = new $className($this->adapter);
             }