+++ testing_migrations/phoenix/20150908065215_add_data.php
@@ -9,33 +9,28 @@ class AddData extends AbstractMigration
 {
     public function up(): void
     {
-//        $this->insert('table_1', [
-//            'id' => 1,
-//            'title' => 'First item',
-//            'alias' => 'first-item',
-//            'is_active' => false,
-//            'self_fk' => null,
-//        ]);
-//
-//        $this->insert('table_1', [
-//            [
-//                'id' => 2,
-//                'title' => 'Second item',
-//                'alias' => 'second-item',
-//                'self_fk' => 1,
-//            ],
-//            [
-//                'id' => 3,
-//                'title' => 'Third item',
-//                'alias' => 'third-item',
-//                'self_fk' => 2,
-//            ]
-//        ]);
+        $this->insert('table_1', [
+            'id' => 1,
+            'title' => 'First item',
+            'alias' => 'first-item',
+            'is_active' => false,
+            'self_fk' => null,
+        ]);
 
-        // temporary changed to plain queries because of failing tests on travis - not sure the reason
-        $this->execute("INSERT INTO table_1 (id, title, alias, is_active, self_fk) VALUES (1, 'First item', 'first-item', false, null);");
-        $this->execute("INSERT INTO table_1 (id, title, alias, self_fk) VALUES (2, 'Second item', 'second-item', 1);");
-        $this->execute("INSERT INTO table_1 (id, title, alias, self_fk) VALUES (3, 'Third item', 'third-item', 2);");
+        $this->insert('table_1', [
+            [
+                'id' => 2,
+                'title' => 'Second item',
+                'alias' => 'second-item',
+                'self_fk' => 1,
+            ],
+            [
+                'id' => 3,
+                'title' => 'Third item',
+                'alias' => 'third-item',
+                'self_fk' => 2,
+            ]
+        ]);
 
         $this->insert('table_2', [
             'id' => 1,
@@ -116,7 +111,7 @@ class AddData extends AbstractMigration
             'col_double' => 3.1415,
             'col_decimal' => 3.1415,
             'col_numeric' => 3.1415,
-            'col_boolean' => true,
+            'col_boolean' => false,
             'col_datetime' => new DateTime(),
             'col_date' => (new DateTime())->format('Y-m-d'),
             'col_time' => '20:30:40',