+++ tests/Database/Adapter/AdapterFactoryTest.php
@@ -6,6 +6,8 @@ use Phoenix\Config\EnvironmentConfig;
 use Phoenix\Database\Adapter\AdapterFactory;
 use Phoenix\Database\Adapter\MysqlAdapter;
 use Phoenix\Database\Adapter\PgsqlAdapter;
+use Phoenix\Database\QueryBuilder\MysqlQueryBuilder;
+use Phoenix\Database\QueryBuilder\MysqlWithJsonQueryBuilder;
 use Phoenix\Exception\InvalidArgumentValueException;
 use PHPUnit\Framework\TestCase;
 
@@ -17,7 +19,59 @@ class AdapterFactoryTest extends TestCase
             'adapter' => 'mysql',
             'dsn' => 'sqlite::memory:',
         ]);
-        $this->assertInstanceOf(MysqlAdapter::class, AdapterFactory::instance($config));
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(MysqlAdapter::class, $adapter);
+        $this->assertInstanceOf(MysqlQueryBuilder::class, $adapter->getQueryBuilder());
+    }
+
+    public function testMysqlVersion550()
+    {
+        $config = new EnvironmentConfig([
+            'adapter' => 'mysql',
+            'version' => '5.5.0',
+            'dsn' => 'sqlite::memory:',
+        ]);
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(MysqlAdapter::class, $adapter);
+        $this->assertInstanceOf(MysqlQueryBuilder::class, $adapter->getQueryBuilder());
+    }
+
+    public function testMysqlVersion508()
+    {
+        $config = new EnvironmentConfig([
+            'adapter' => 'mysql',
+            'version' => '5.0.8',
+            'dsn' => 'sqlite::memory:',
+        ]);
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(MysqlAdapter::class, $adapter);
+        $queryBuilder = $adapter->getQueryBuilder();
+        $this->assertInstanceOf(MysqlQueryBuilder::class, $queryBuilder);
+        $this->assertNotInstanceOf(MysqlWithJsonQueryBuilder::class, $queryBuilder);
+    }
+
+    public function testMysqlVersion578()
+    {
+        $config = new EnvironmentConfig([
+            'adapter' => 'mysql',
+            'version' => '5.7.8',
+            'dsn' => 'sqlite::memory:',
+        ]);
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(MysqlAdapter::class, $adapter);
+        $this->assertInstanceOf(MysqlWithJsonQueryBuilder::class, $adapter->getQueryBuilder());
+    }
+
+    public function testMysqlVersion8019()
+    {
+        $config = new EnvironmentConfig([
+            'adapter' => 'mysql',
+            'version' => '8.0.19',
+            'dsn' => 'sqlite::memory:',
+        ]);
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(MysqlAdapter::class, $adapter);
+        $this->assertInstanceOf(MysqlWithJsonQueryBuilder::class, $adapter->getQueryBuilder());
     }
 
     public function testPgsql()
@@ -26,7 +80,8 @@ class AdapterFactoryTest extends TestCase
             'adapter' => 'pgsql',
             'dsn' => 'sqlite::memory:',
         ]);
-        $this->assertInstanceOf(PgsqlAdapter::class, AdapterFactory::instance($config));
+        $adapter = AdapterFactory::instance($config);
+        $this->assertInstanceOf(PgsqlAdapter::class, $adapter);
     }
 
     public function testUnknown()